+++ test/landrush/action/setup_test.rb
@@ -5,8 +5,13 @@ require 'landrush/action/setup'
 module Landrush
   module Action
     describe Setup do
-      before { ResolverConfig.sudo = ''     }
-      after  { ResolverConfig.sudo = 'sudo' }
+      before do
+        ResolverConfig.sudo = ''
+      end
+
+      after do
+        ResolverConfig.sudo = 'sudo'
+      end
 
       it "calls the next app in the chain" do
         env = fake_environment