+++ test/support/delete_fake_working_dir.rb
@@ -1,0 +1,12 @@
+module DeleteFakeWorkingDirHooks
+  def teardown
+    super
+    Landrush::Server.working_dir.rmtree if Landrush::Server.working_dir.directory?
+  end
+end
+
+module MiniTest
+  class Spec
+    include DeleteFakeWorkingDirHooks
+  end
+end